Use Arrow to define ActionFunction
As suggested by @sadache.

This is currently broken, as it removes the possible parameterization at
ActionBuilder.apply time over the content type (forcing AnyContent by default).
This is incompatible with many current BodyParsers.  Because the Request type
has to be concrete to define the Arrow instance, I see no way around this.
Pending more discussion on #1986.